[bir][birgen] Separate out autogenerated files
