+++ bir/birgen2/BIRGen2.cpp
@@ -6,4 +6,18 @@
 #include "belalang/BIRGen/BIRGen.h"
 
 #include "BIRGen2.h"
+
+namespace belalang {
+namespace birgen2 {
+
+std::unique_ptr<BIRGen2> create_birgen2(uintptr_t gen_ptr) {
+  auto *gen = reinterpret_cast<birgen::BIRGen *>(gen_ptr);
+  return std::make_unique<BIRGen2>(*gen);
+}
+
+BIRGen2::BIRGen2(::belalang::birgen::BIRGen &gen) : gen(gen) {}
+
 #include "Bindings.cpp.inc"
+
+} // birgen2
+} // belalang
